+++ lib/templates/bomboniere/src/App.vue
@@ -52,11 +52,6 @@ h1,h2,h3,h4,h5,h6,p {
   margin: 2px 0px;
   padding: 0px 8px;
   height: 20px;
-}
-
-.prefix-name {
-  width: 300px;
-  float: left;
   text-align: left;
 }
 
@@ -78,5 +73,30 @@ h1,h2,h3,h4,h5,h6,p {
   width: 400px;
   float: left;
   text-align: left;
+}
+
+/* accordion */
+.accordion-item{
+  position: relative;
+}
+.accordion-item-head{
+  cursor: pointer;
+}
+.accordion-item-head:after{
+  content: ' > ';
+  display: block;
+  position: absolute;
+  right: 25px;
+  transform: rotate(90deg) scaleY(2);
+  top: 0px;
+}
+.accordion-item-head.active:after{
+  content: ' < ';
+}
+.accordion-item-body{
+  display: none;
+}
+.accordion-item-head.active + .accordion-item-body{
+  display: block !important;
 }
 </style>