+++ lib/templates/bomboniere/src/components/HelloWorld.vue
@@ -44,7 +44,7 @@
                   <div class="method">{{ action_details.method }}</div>
                   <div class="path">{{ action_details.path }}</div>
                   <div v-for="responses_details in action_details.responses.responses_details">
-                    <div class="responses_details">{{ responses_details.combinations.combinations_cover_percent }}
+                    <div class="responses_details">{{ responses_details.combinations.cover_percent }}
                       {{ responses_details.method }}
                     </div>
                   </div>