+++ lib/templates/bomboniere/src/views/Action.vue
@@ -11,12 +11,14 @@
             </div>
 
             <div v-if="action.responses.tests_without_responses.length != 0">
-              <div class="tests_without_responses">tests without responses:
-                {{ action.responses.tests_without_responses.length }}
-              </div>
-              <div v-for="test_without_responses in action.responses.tests_without_responses">
-                <div class="test_without_responses">
-                  {{ test_without_responses }}
+              <div class="accordion-item">
+                <div class="tests_without_responses accordion-item-head" v-on:click="accordion">
+                  tests without responses: {{ action.responses.tests_without_responses.length }} ✖
+                </div>
+                <div class="accordion-item-body">
+                  <div v-for="test_without_responses in action.responses.tests_without_responses">
+                    <div class="test_without_responses">{{ test_without_responses }} ✖</div>
+                  </div>
                 </div>
               </div>
             </div>