+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * add responses join in new report [#43](https://github.com/funbox/fitting/issues/43)
   * add combinations join in new report [#47](https://github.com/funbox/fitting/issues/47)
   * add action page [#49](https://github.com/funbox/fitting/issues/49)
+  * show more information in new report [#51](https://github.com/funbox/fitting/issues/51)
 
 ### 2.13.1 - 2020-04-17