+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * move json-schemas to separate files [#57](https://github.com/funbox/fitting/issues/57)
 * fixes
   * fix vulnerability CVE-2020-7660 [#53](https://github.com/funbox/fitting/issues/53)
+  * fix cover if response without combinations [#59](https://github.com/funbox/fitting/issues/59)
 
 ### 2.13.1 - 2020-04-17