+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * move json-schemas to separate files [#57](https://github.com/funbox/fitting/issues/57)
   * add method for tests without actions [#61](https://github.com/funbox/fitting/issues/61)
   * show tests without responses [#63](https://github.com/funbox/fitting/issues/63)
+  * add console output for new report [#69](https://github.com/funbox/fitting/issues/69)
 * fixes
   * fix vulnerability CVE-2020-7660 [#53](https://github.com/funbox/fitting/issues/53)
   * fix cover if response without combinations [#59](https://github.com/funbox/fitting/issues/59)