+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * add method for tests without actions [#61](https://github.com/funbox/fitting/issues/61)
   * show tests without responses [#63](https://github.com/funbox/fitting/issues/63)
   * add console output for new report [#69](https://github.com/funbox/fitting/issues/69)
+  * check test in new console [#71](https://github.com/funbox/fitting/issues/71)
 * fixes
   * fix vulnerability CVE-2020-7660 [#53](https://github.com/funbox/fitting/issues/53)
   * fix cover if response without combinations [#59](https://github.com/funbox/fitting/issues/59)